refactor(oh7): split model tier resolution into core, keep LinkAgent remap in linker (#252)
