+++ src/cli/link.rs
@@ -72,7 +72,7 @@ pub async fn execute(
 
     // 2b. Resolve deprecated model aliases before remapping
     for agent in &mut link_agents {
-        crate::linker::model_aliases::resolve_model_deprecations(
+        crate::core::model_aliases::resolve_model_deprecations(
             &mut agent.model,
             &mut agent.model_fallback,
         );
