+++ src/core/mod.rs
@@ -4,6 +4,8 @@ pub mod config;
 pub mod dependency_resolver;
 pub(crate) mod embedded;
 pub mod events;
+pub mod model_aliases;
+pub mod model_resolution;
 pub mod model_updater;
 #[allow(dead_code)]
 pub mod orchestration;
