+++ src/linker/mod.rs
@@ -2,7 +2,6 @@ mod claude;
 mod codex;
 mod copilot;
 mod gemini;
-pub mod model_aliases;
 pub mod model_resolution;
 mod opencode;
